Refactor further to avoid merge conflicts
This change moves custom code from sceptre.cli.template into
sceptre.cli.custom and custom unit tests to tests/test_actions_custom.py
and tests/test_cli_custom.py. Also creates classes SceptrePlanCustom and
StackActionsCustom for the same reason. These are inherited and extended
by SceptrePlan and StackActions respectively.

The intention is to avoid the risk of merge conflicts and make merge
conflicts that do occur easier to resolve.
